+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -236,5 +236,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-30T12:03:32.471Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-30T12:45:57.311Z")
     r"""The date and time the resource was last updated"""