+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -571,7 +571,7 @@ class HTTPTool(BaseModel):
     http: HTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6D83JNJWC7MMM2HDS4PNB5M"
+        "01K6DAH8FFB81QN5N50T5P5219"
     )
 
     requires_approval: Optional[bool] = False