+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -603,7 +603,7 @@ class RunAgentRequestToolHTTPTool(BaseModel):
     http: RunAgentRequestToolHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6D83JNPHYMAVD3B8WC12Z76"
+        "01K6DAH8FK2Z1T6C8NZXPXTJR5"
     )
 
     requires_approval: Optional[bool] = False